+++ src/controllers/OrdersController.php
@@ -35,7 +35,6 @@ use craft\commerce\web\assets\commerceui\CommerceOrderAsset;
 use craft\db\Query;
 use craft\db\Table as CraftTable;
 use craft\elements\Address;
-use craft\elements\db\AddressQuery;
 use craft\elements\User;
 use craft\errors\ElementNotFoundException;
 use craft\errors\InvalidElementException;
@@ -218,7 +217,7 @@ class OrdersController extends Controller
 
         $alreadyCompleted = $order->isCompleted;
         // Set data from request to the order
-        $this->_updateOrder($order, $orderRequestData);
+        $this->_updateOrder($order, $orderRequestData, false);
         $markAsComplete = !$alreadyCompleted && $order->isCompleted;
 
         // We don't want to save it as completed yet since we will markAsComplete() after saving the cart
@@ -643,11 +642,10 @@ class OrdersController extends Controller
             return $this->asFailure(message: Craft::t('commerce', 'Order not found.'));
         }
 
-        /** @var AddressQuery $addressQuery */
-        $addressQuery = Address::find()
-            ->id($addressId);
-        $address = $addressQuery
+        /** @var Address|null $address */
+        $address = Address::find()
             ->ownerId($order->id)
+            ->id($addressId)
             ->one();
 
         if (!$address) {
@@ -851,7 +849,7 @@ class OrdersController extends Controller
         /** @var Gateway $gateway */
         foreach ($gateways as $key => $gateway) {
             // If gateway adapter does no support backend cp payments.
-            if (!$gateway->cpPaymentsEnabled() || $gateway instanceof MissingGateway) {
+            if ($gateway->availableForUseWithOrder($order) === false || !$gateway->cpPaymentsEnabled() || $gateway instanceof MissingGateway) {
                 unset($gateways[$key]);
                 continue;
             }
@@ -1223,13 +1221,15 @@ class OrdersController extends Controller
      * @throws InvalidElementException
      * @throws UnsupportedSiteException
      */
-    private function _updateOrder(Order $order, $orderRequestData): void
+    private function _updateOrder(Order $order, $orderRequestData, bool $tryAutoSet = true): void
     {
         $order->setRecalculationMode($orderRequestData['order']['recalculationMode']);
         $order->reference = $orderRequestData['order']['reference'];
 
+        $hasSetCustomer = false;
         $customerId = $orderRequestData['order']['customerId'] ?? null;
         if ($customerId && $customer = Craft::$app->getUsers()->getUserById($customerId)) {
+            $hasSetCustomer = true;
             $order->setCustomer($customer);
         } else {
             $order->setCustomer();
@@ -1242,33 +1242,46 @@ class OrdersController extends Controller
         $order->shippingMethodHandle = $orderRequestData['order']['shippingMethodHandle'];
         $order->suppressEmails = $orderRequestData['order']['suppressEmails'] ?? false;
 
-        $getAddress = static function($address, $orderId, $title) {
-            if ($address && ($address['id'] && ($address['ownerId'] != $orderId || isset($address['_copy'])))) {
-                if (isset($address['_copy'])) {
-                    unset($address['_copy']);
-                }
-                $address = Craft::$app->getElements()->getElementById($address['id'], Address::class);
-                $address = Craft::$app->getElements()->duplicateElement($address, ['ownerId' => $orderId, 'title' => $title]);
-            } elseif ($address && ($address['id'] && $address['ownerId'] == $orderId)) {
-                /** @var AddressQuery $addressQuery */
-                $addressQuery = Address::find()->id($address['id']);
-                $address = $addressQuery->ownerId($address['ownerId'])->one();
+        $submittedBillingAddress = $orderRequestData['order']['billingAddress'] ?? null;
+        $submittedShippingAddress = $orderRequestData['order']['shippingAddress'] ?? null;
+
+        if ($tryAutoSet && $hasSetCustomer && $submittedShippingAddress === null && $submittedBillingAddress === null) {
+            // Try and auto set addresses if the customer has changed and no address data is submitted
+            // Remove any lingering addresses from previous saves
+            if (!$order->isCompleted) {
+                $order->setBillingAddress(null);
+                $order->setShippingAddress(null);
             }
 
-            return $address;
-        };
-        $billingAddress = $getAddress($orderRequestData['order']['billingAddress'] ?? null, $orderRequestData['order']['id'], Craft::t('commerce', 'Billing Address'));
-        $order->setBillingAddress($billingAddress);
+            $order->autoSetAddresses();
+        } else {
+            $getAddress = static function($address, $orderId, $title) {
+                if ($address && ($address['id'] && ($address['ownerId'] != $orderId || isset($address['_copy'])))) {
+                    if (isset($address['_copy'])) {
+                        unset($address['_copy']);
+                    }
+                    $address = Craft::$app->getElements()->getElementById($address['id'], Address::class);
+                    $address = Craft::$app->getElements()->duplicateElement($address, ['ownerId' => $orderId, 'title' => $title]);
+                } elseif ($address && ($address['id'] && $address['ownerId'] == $orderId)) {
+                    /** @var Address|null $address */
+                    $address = Address::find()->ownerId($address['ownerId'])->id($address['id'])->one();
+                }
 
-        $shippingAddress = $getAddress($orderRequestData['order']['shippingAddress'] ?? null, $orderRequestData['order']['id'], Craft::t('commerce', 'Shipping Address'));
-        $order->setShippingAddress($shippingAddress);
+                return $address;
+            };
+            $billingAddress = $getAddress($submittedBillingAddress, $orderRequestData['order']['id'], Craft::t('commerce', 'Billing Address'));
+            $order->setBillingAddress($billingAddress);
 
-        if (isset($orderRequestData['order']['sourceBillingAddressId'])) {
-            $order->sourceBillingAddressId = $orderRequestData['order']['sourceBillingAddressId'];
-        }
+            $shippingAddress = $getAddress($submittedShippingAddress, $orderRequestData['order']['id'], Craft::t('commerce', 'Shipping Address'));
+            $order->setShippingAddress($shippingAddress);
 
-        if (isset($orderRequestData['order']['sourceShippingAddressId'])) {
-            $order->sourceShippingAddressId = $orderRequestData['order']['sourceShippingAddressId'];
+            if (isset($orderRequestData['order']['sourceBillingAddressId'])) {
+                $order->sourceBillingAddressId = $orderRequestData['order']['sourceBillingAddressId'];
+            }
+
+            if (isset($orderRequestData['order']['sourceShippingAddressId'])) {
+                $order->sourceShippingAddressId = $orderRequestData['order']['sourceShippingAddressId'];
+            }
         }
 
         $shippingMethod = $order->shippingMethodHandle ? Plugin::getInstance()->getShippingMethods()->getShippingMethodByHandle($order->shippingMethodHandle) : null;
@@ -1527,6 +1540,7 @@ class OrdersController extends Controller
                 $row['newLineItemOptionsSignature'] = LineItem::generateOptionsSignature([]);
                 $row['description'] = Html::encode($row['description']);
                 $row['sku'] = Html::encode($row['sku']);
+                $row['qty'] = '';
                 $purchasables[] = $row;
             }
         }