+++ src/services/Discounts.php
@@ -255,7 +255,7 @@ class Discounts extends Component
         } else {
             // We use a round the time so we can have a cache within the same request (rounded to 1 minute flat, no seconds)
             $date = new DateTime();
-            $date->setTime($date->format('H'), round($date->format('i') / 1) * 1);
+            $date->setTime((int)$date->format('H'), (int)(round($date->format('i') / 1) * 1));
         }
 
         // Coupon condition key
@@ -666,6 +666,7 @@ class Discounts extends Component
         $record->maxPurchaseQty = $model->maxPurchaseQty;
         $record->baseDiscount = $model->baseDiscount;
         $record->baseDiscountType = $model->baseDiscountType;
+        $record->purchaseTotal = $model->purchaseTotal;
         $record->perItemDiscount = $model->perItemDiscount;
         $record->percentDiscount = $model->percentDiscount;
         $record->percentageOffSubject = $model->percentageOffSubject;
@@ -1167,6 +1168,7 @@ class Discounts extends Component
                 '[[discounts.perEmailLimit]]',
                 '[[discounts.perItemDiscount]]',
                 '[[discounts.perUserLimit]]',
+                '[[discounts.purchaseTotal]]',
                 '[[discounts.purchaseQty]]',
                 '[[discounts.sortOrder]]',
                 '[[discounts.stopProcessing]]',