+++ src/web/assets/commerceui/src/js/order/components/InputError.vue
@@ -1,10 +1,7 @@
 <template>
     <div>
-        <div
-            v-if="draft.order.errors && draft.order.errors[errorKey]"
-            class="error"
-        >
-            {{ draft.order.errors[errorKey] }}
+        <div v-if="draft.errors && draft.errors[errorKey]" class="error">
+            {{ draft.errors[errorKey] }}
         </div>
     </div>
 </template>