+++ src/widgets/TotalOrdersByCountry.php
@@ -9,13 +9,13 @@ namespace craft\commerce\widgets;
 
 use Craft;
 use craft\base\Widget;
+use craft\commerce\base\StatWidgetTrait;
 use craft\commerce\stats\TotalOrdersByCountry as TotalOrdersByCountryStat;
 use craft\commerce\web\assets\statwidgets\StatWidgetsAsset;
 use craft\helpers\ArrayHelper;
 use craft\helpers\DateTimeHelper;
 use craft\helpers\Html;
 use craft\helpers\StringHelper;
-use DateTime;
 
 /**
  * Total Orders By Country widget
@@ -29,23 +29,10 @@ use DateTime;
  */
 class TotalOrdersByCountry extends Widget
 {
-    /**
-     * @var int|DateTime|null
-     */
-    public mixed $startDate = null;
+    use StatWidgetTrait;
 
     /**
-     * @var int|DateTime|null
-     */
-    public mixed $endDate = null;
-
-    /**
-     * @var string|null
-     */
-    public ?string $dateRange = null;
-
-    /**
-     * @var string Options 'billing', 'shippinh'.
+     * @var string Options 'billing', 'shipping'.
      */
     public string $type;
 
@@ -91,6 +78,10 @@ class TotalOrdersByCountry extends Widget
             DateTimeHelper::toDateTime($this->startDate, true),
             DateTimeHelper::toDateTime($this->endDate, true)
         );
+
+        if (!empty($this->orderStatuses)) {
+            $this->_stat->setOrderStatuses($this->orderStatuses);
+        }
     }
 
     /**
@@ -174,6 +165,7 @@ class TotalOrdersByCountry extends Widget
         return Craft::$app->getView()->renderTemplate('commerce/_components/widgets/orders/country/settings', [
             'id' => $id,
             'namespaceId' => $namespaceId,
+            'orderStatuses' => $this->getOrderStatusOptions(),
             'widget' => $this,
             'typeOptions' => $this->_typeOptions,
         ]);